[Refactor/#107 옷 차림 코드개선]
