+++ weady/weady/Sources/Feature/Clothes/Views/ClothingRecommendationView.swift
@@ -17,9 +17,10 @@ struct ClothingRecommendationView: View {
     //private enum Route: Hashable { case weadyboard }
     @Environment(HomeRouter.self) private var router
     
-    private enum HelpStep: Hashable { case intro, details }
-    @State private var showHelp = false //HelpGuideCardView
-    @State private var helpStep: HelpStep = .intro //HelpGuideCardView 1,2
+    enum HelpStep: Hashable { case intro, details }
+    @State private var showHelp = false
+    @State private var helpStep: HelpStep = .intro
+    
     
     let same = Date()
     // 기본 init: 내부에서 VM 생성
@@ -38,205 +39,131 @@ struct ClothingRecommendationView: View {
     }
     
     var body: some View {
-
-            ZStack{
-                Image("backgroundImage")
-                    .resizable()
-                    .scaledToFill()
-                    .ignoresSafeArea()
-                
-                VStack(alignment: .leading, spacing: 0) {
-                    // 주소
-                    HStack {
-                        Image("mapIcon")
-                            .resizable()
-                            .frame(width: 12, height: 17)
-                        Text(vm.addressText)
-                            .fontName(.bodySemibold16)
-                            .foregroundStyle(.appwhite100)
-                        //지역 선택 화면으로 가는 버튼
-                        Button {
-                            router.push(.weatherlocation)
-                        } label: {
-                            Image("clothesDownIcon")
-                                .resizable()
-                                .frame(width: 10, height: 4)
-                                .contentShape(Rectangle())
-                        }
-                        .buttonStyle(.plain)
-                        
-                        Spacer()
-                    }
-                    .padding(.horizontal, 120)
-                    .padding(.top, 41)
-                    
-                    // 추천 옷 이미지
-#if DEBUG
-                    // Preview에서는 네트워크 호출 없이 바로 에셋 이미지를 보여주기
-                    Image("teeShirt")
+        
+        ZStack{
+            Image("backgroundImage")
+                .resizable()
+                .scaledToFill()
+                .ignoresSafeArea()
+            
+            VStack(alignment: .leading, spacing: 0) {
+                // 주소
+                HStack {
+                    Image("mapIcon")
                         .resizable()
-                        .frame(width: 191, height: 173)
-                        .frame(maxWidth: .infinity, alignment: .center)
-#else
-                    AsyncImage(url: vm.clothingImageUrl) { image in
-                        image.resizable().aspectRatio(contentMode: .fit)
-                    } placeholder: {
-                        ProgressView()
+                        .frame(width: 12, height: 17)
+                    Text(vm.addressText)
+                        .fontName(.bodySemibold16)
+                        .foregroundStyle(.appwhite100)
+                    //지역 선택 화면으로 가는 버튼
+                    Button {
+                        router.push(.weatherlocation)
+                    } label: {
+                        Image("clothesDownIcon")
+                            .resizable()
+                            .frame(width: 10, height: 4)
+                            .contentShape(Rectangle())
                     }
+                    .buttonStyle(.plain)
+                    
+                    Spacer()
+                }
+                .padding(.horizontal, 120)
+                .padding(.top, 41)
+                
+                // 추천 옷 이미지
+#if DEBUG
+                // Preview에서는 네트워크 호출 없이 바로 에셋 이미지를 보여주기
+                Image("teeShirt")
+                    .resizable()
                     .frame(width: 191, height: 173)
                     .frame(maxWidth: .infinity, alignment: .center)
+#else
+                AsyncImage(url: vm.clothingImageUrl) { image in
+                    image.resizable().aspectRatio(contentMode: .fit)
+                } placeholder: {
+                    ProgressView()
+                }
+                .frame(width: 191, height: 173)
+                .frame(maxWidth: .infinity, alignment: .center)
 #endif
-                    
-                    // 추천 문구
-                    VStack(){
-                        HStack(spacing: 0){
-                            Text("오늘은 ")
-                                .fontName(.titleMedium24)
-                            Text(vm.clothingName)
-                                .fontName(.titleBold24)
-                            Text(vm.subjectParticle)
-                                .fontName(.titleMedium24)
-                        }
-                        Text("딱 좋은 날이에요.")
+                
+                // 추천 문구
+                VStack(){
+                    HStack(spacing: 0){
+                        Text("오늘은 ")
+                            .fontName(.titleMedium24)
+                        Text(vm.clothingName)
+                            .fontName(.titleBold24)
+                        Text(vm.subjectParticle)
                             .fontName(.titleMedium24)
                     }
-                    .padding(.top, 1)
-                    .frame(maxWidth: .infinity, alignment: .center)
-                    .foregroundStyle(.appwhite100)
-                    
-                    // 체감온도
-                    HStack(spacing: 9) {
-                        Image("thermometer")
+                    Text("딱 좋은 날이에요.")
+                        .fontName(.titleMedium24)
+                }
+                .padding(.top, 1)
+                .frame(maxWidth: .infinity, alignment: .center)
+                .foregroundStyle(.appwhite100)
+                
+                // 체감온도
+                HStack(spacing: 9) {
+                    Image("thermometer")
+                        .resizable()
+                        .frame(width: 8, height: 16)
+                    Text("체감 \(vm.feelTemp)°")
+                        .fontName(.captionRegular14)
+                        .foregroundStyle(.appwhite100)
+                }
+                .padding(.horizontal, 165)
+                .padding(.top, 14)
+                
+                Spacer().frame(height: 21)
+                
+                // 기온 차트
+                TemperatureChartView(chartItems: vm.chartItems)
+                    .frame(width: 337, height: 138)
+                    .padding(.top, 21)
+                    .padding(.horizontal, 20)
+                
+                // 체감온도 기준
+                Button {
+                    withAnimation(.easeInOut(duration: 0.2)) {
+                        helpStep = .intro   // 첫 화면부터 시작
+                        showHelp = true     // 오버레이 열기
+                    }
+                } label: {
+                    HStack(spacing: 3) {
+                        Image("helpIcon")
                             .resizable()
-                            .frame(width: 8, height: 16)
-                        Text("체감 \(vm.feelTemp)°")
-                            .fontName(.captionRegular14)
-                            .foregroundStyle(.appwhite100)
+                            .frame(width: 12, height: 12)
+                        Text("체감온도 기준")
+                            .fontName(.metaMedium8)
+                            .foregroundColor(.appwhite100)
                     }
-                    .padding(.horizontal, 165)
-                    .padding(.top, 14)
-                    
-                    Spacer().frame(height: 21)
-                    
-                    // 기온 차트
-                    TemperatureChartView(chartItems: vm.chartItems)
-                        .frame(width: 337, height: 138)
-                        .padding(.top, 21)
-                        .padding(.horizontal, 20)
-                    
-                    // 체감온도 기준
-                    Button {
-                        withAnimation(.easeInOut(duration: 0.2)) {
-                            helpStep = .intro   // 첫 화면부터 시작
-                            showHelp = true     // 오버레이 열기
-                        }
-                    } label: {
-                        HStack(spacing: 3) {
-                            Image("helpIcon")
-                                .resizable()
-                                .frame(width: 12, height: 12)
-                            Text("체감온도 기준")
-                                .fontName(.metaMedium8)
-                                .foregroundColor(.appwhite100)
-                        }
-                    }
-                    .padding(.leading, 312)
-                    .padding(.top, 13)
-                    
-                    Button {
-                            router.push(.weadyboard)
-                                   } label: {
-                                       NavigationRowLabel(
-                                           title: "다른 사람들은 어떻게 입었는지 보러가기",
-                                           images: ["howPic1","howPic2","howPic3"]
-                                      )
-                                       .contentShape(RoundedRectangle(cornerRadius: 10))
-                                   }
-                    .frame(maxWidth: .infinity, alignment: .center)
-                    .padding(.top, 13)
-                    .padding(.bottom, 19)
                 }
-                if showHelp {
-                    ZStack {
-                        Color.black.opacity(0.45)
-                            .ignoresSafeArea()
-                            .onTapGesture { withAnimation(.easeOut(duration: 0.2)) { showHelp = false } }
-
-                        if helpStep == .intro {
-                            HelpGuideCardView(
-                                onClose: { withAnimation(.easeOut(duration: 0.2)) { showHelp = false } },
-                                onNext:  { withAnimation(.easeInOut) { helpStep = .details } }
-                            )
-                            .transition(.move(edge: .trailing).combined(with: .opacity))
-                        } else {
-                            HelpGuideCardView2(
-                                onClose: { withAnimation(.easeOut(duration: 0.2)) { showHelp = false } },
-                                onBack:  { withAnimation(.easeInOut) { helpStep = .intro } }
-                            )
-                            .transition(.move(edge: .leading).combined(with: .opacity))
-                        }
-                    }
-                    .zIndex(3)
-                    
-                }
-                    
+                .padding(.leading, 312)
+                .padding(.top, 13)
+                
+                WeadyboardCTA(
+                    title: "다른 사람들은 어떻게 입었는지 보러가기",
+                    images: ["howPic1","howPic2","howPic3"],
+                    onTap: { router.push(.weadyboard) }
+                )
+                .frame(maxWidth: .infinity, alignment: .center)
+                .padding(.top, 13)
+                .padding(.bottom, 19)
             }
+            
+            // 도움말 오버레이
+            HelpOverlay(
+                isPresented: $showHelp,
+                step: $helpStep,
+                onClose: { withAnimation(.easeOut(duration: 0.2)) { showHelp = false }}
+            )
         }
-    }
-
-
-struct NavigationRowLabel: View {
-    let title: String
-    let images: [String]
-    var body: some View {
-        HStack(spacing: 0) {
-            Text(title)
-                .fontName(.captionMedium14)
-                .foregroundColor(.black100)
-                .padding(.trailing, 15)
-            OverlappingThumbnails(images: images, size: 28, overlap: 14)
-                .padding(.trailing, 22)
-            Image("clothesRightIcon")
-                .resizable()
-                .frame(width: 23, height: 23)
-        }
-        .padding(.leading, 30)
-        .padding(.trailing, 10)
-        .padding(.vertical, 12)
-        .background(
-            RoundedRectangle(cornerRadius: 10).fill(Color.appwhite100)
-        )
     }
 }
-
-// 3장까지 겹치는 썸네일
-struct OverlappingThumbnails: View {
-    let images: [String]
-    var size: CGFloat = 28
-    let aspect: CGFloat = 21.0/28.0
-    var overlap: CGFloat = 14     // 겹치는 정도
-    var centerLift: CGFloat = 2   // 가운데 이미지 위로 올리는 정도
-    var sideDrop: CGFloat  = 4    // 양쪽 이미지 아래로 내리는 정도
-
-    var body: some View {
-        let capped = Array(images.prefix(3))
-        let mid = capped.count / 2
-
-        HStack(spacing: -overlap) {
-            ForEach(capped.indices, id: \.self) { i in
-                Image(capped[i])
-                    .resizable()
-                    .scaledToFill()
-                    .frame(width: size * aspect, height: size)
-                    .rotationEffect(.degrees(i == mid ? 0 : (i < mid ? -6 : 6)))
-                    .offset(y: i == mid ? centerLift : sideDrop)
-                    .zIndex(Double(i))
-            }
-        }
-        .frame(height: size)
-        .offset(y: -((sideDrop + centerLift + sideDrop) / 3))
-    }
-}
+        
 
 // MARK: - Preview & Mock
 extension ClothingRecommendationViewModel {
